moved recursive_get_result
